+++ src/utils/index.js
@@ -26,6 +26,7 @@ exports["speedometer-runtime"] = require("./speedometer-runtime.js");
 exports["debug-runtime"] = require("./debug-runtime.js");
 exports["sensitive-fields"] = require("./sensitive-fields.js");
 exports["message-guard"] = require("./message-guard.js");
+exports["path-guard"] = require("./path-guard.js");
 exports["config-security"] = require("./config-security.js");
 exports["ssh-credential-fields"] = require("./ssh-credential-fields.js");
 exports["ftp-credential-fields"] = require("./ftp-credential-fields.js");